feat: move BookmarkType, IExpiration, IZapRaiser, AdditiveComplexFeedFilter to commons
Part of the KMP iOS migration (#2238).

Moves these thin types/interfaces from the amethyst module to commons:
- BookmarkType enum
- IExpiration interface (post creator)
- IZapRaiser interface (zap UI)
- AdditiveComplexFeedFilter abstract class (extends FeedFilter already in commons)

Backward-compatible typealiases left at original locations.
